src/common/exceptions/filters/http-exception.filter.ts
import {
  ExceptionFilter,
  Catch,
  HttpException,
  Logger,
  ArgumentsHost,
} from '@nestjs/common';
import { Response } from 'express';

@Catch(HttpException)
export class HttpExceptionFilter implements ExceptionFilter {
  readonly logger = new Logger(HttpExceptionFilter.name);

  catch(exception: HttpException, host: ArgumentsHost) {
    const context = host.switchToHttp();
    const response = context.getResponse<Response>();

    const status = exception.getStatus();
    this.logger.error(`${exception.message}, status: ${status}`);

    const newResponse = response.status(status)
    newResponse.json({
      statusCode: status,
      timestamp: new Date().toISOString(),
      message: exception.message,
    });
  }
}
